+++ app/src/page.tsx
@@ -112,10 +112,6 @@ const workspaceStore = new WorkspaceStore(
       sessionId,
       requestId
     }),
-    getRequestActiveData: (requestId: string) => core.invoke('get_request_active_data', {
-      sessionId,
-      requestId
-    }),
     add: (entityType: EntityType, relativeToId: string | null, relativePosition: IndexedEntityPosition | null, cloneFromId: string | null) =>
       core.invoke<string>('add', {
         sessionId,